backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-jose
passlib[bcrypt]
pycryptodome
motor
pymongo[srv]
